fix dimensions, fix lightstep example
lightstep was blowing up because the metrics that were being sent
were missing dimension values. This was caused in turn by having
an Any type value field on the Dimension interface. By removing
that harmful method the compiler was able to show the issue &
fix the missing data in the OTLP translation.

Opentelemetry metrics export works now.