work around lightstep histogram issue
Lightstep can't handle a histogram with 1 populated bucket. It
interprets the bucket with data as covering a range from negative
infinity to the bucket value. This is a bug fix recommended by
Lightstep support to allow histograms with 1 value to show up,
but it remains unclear whether there are further issues in
lightstep interpreting sparse histograms.