add manual Metrics lifetime path
Usually metrics should be consumed structurally in Kotlin - like
```
metricsFactory.record("interesting_work") { metrics ->
    // some stuff
}
```

but sometimes you have a workflow that revolves around a finite
state machine or the like. For that kind of workflow, the lifecycle
of a Metrics object may not map well to a stack frame. This change
gives an Internals view of the lifecycle methods the MetricsFactory
uses for those special use cases as an escape hatch.