add ability to send metrics to generic OTLP ingest point (#34)
Lightstep is not the only provider of metrics with an OTLP ingestion URL. This PR adds the ability for consumers to use any OTLP ingestion point for sending metrics.

Maintains backwards compatibility with consumers of original APIs without changing business logic. A lot of the code is shared, so this was DRY'd up to keep things tidy.

Finally, made opentelemetryClient() a public function. This allows consumers to write extension functions local to their own code to support features like bifrocated metrics ingestion (e.g. I want to send my metrics to providers X and Y) without pigeon-holing the goometrics library itself to support this.